rpk: Add 'rpk redpanda check', deprecate 'rpk check'
